Add score of tap result 'Great' in result UI
UI
- New result UI form tap result 'Great'

Misc
- Adjust thickness of font
- Adjust definition of TapResult

+++ DDR/Assets/Scripts/UI/UICombo.cs
@@ -43,7 +43,7 @@ public class UICombo : MonoBehaviour {
     private void OnTapResult(BaseEventArgs args) {
         TapResultEventArgs trArgs = args as TapResultEventArgs;
 
-        if (trArgs.TR != (int) TapResult.Miss) {
+        if (trArgs.TR != TapResult.Miss) {
             _comboCount += 1;
         }
         else {
